+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
   - `ZoweVsCodeExtension.showVsCodeMessage` - use `Gui.showMessage` instead.
   - `ZoweVsCodeExtension.inputBox` - use `Gui.showInputBox` instead.
   - `ZoweVsCodeExtension.promptCredentials` - use `ZoweVsCodeExtension.updateCredentials` instead.
+- Added an optional function `openWithEncoding` to the `IZoweTree` interface to open file using encoding specified by user. [#2931](https://github.com/zowe/zowe-explorer-vscode/pull/2931)
 
 ### Bug fixes